backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
channels>=4.0,<5.0
daphne>=4.0,<5.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<5.0
